pyproject.toml
[project]
name = "mnemonify"
version = "0.1.0"
description = "Generate mnemonics from kanji"
authors = [
    {name = "alexnachyn"}
]
readme = "README.md"
requires-python = ">=3.11,<=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "jaconv (>=0.4.0,<0.5.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "transformers (>=4.55.0,<5.0.0)",
    "peft (>=0.17.0,<0.18.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "trl (>=0.21.0,<0.22.0)",
    "bitsandbytes (>=0.46.1,<0.47.0)",
    "torch (>=2.8.0,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "tensorboardx (>=2.6.4,<3.0.0)",
    "llama-cpp-python (>=0.3.15,<0.4.0)",
    "mistral-common (>=1.8.3,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
